feat(44611022): plugin (#210)

+++ superset-frontend/src/dashboard/actions/nativeFilters.ts
@@ -1,5 +1,6 @@
 // DODO was here
 import {
+  API_HANDLER, // DODO added 44611022
   FilterConfiguration,
   Filters,
   FilterSet, // DODO added 44211751
@@ -16,6 +17,8 @@ import { FilterSetFullData } from 'src/DodoExtensions/FilterSets/types'; // DODO
 import { HYDRATE_DASHBOARD } from './hydrate';
 import { dashboardInfoChanged } from './dashboardInfo';
 import { DashboardInfo, RootState } from '../types';
+
+const isStandalone = process.env.type === undefined; // DODO added 44611022
 
 export const SET_FILTER_CONFIG_BEGIN = 'SET_FILTER_CONFIG_BEGIN';
 export interface SetFilterConfigBegin {
@@ -274,24 +277,27 @@ export interface UpdateFilterSetFail {
 }
 
 const getFilterSets = (dashboardId: number) => async (dispatch: Dispatch) => {
-  const fetchFilterSets = async () =>
-    // if (process.env.type === undefined) {
-    makeApi<
-      null,
-      {
-        count: number;
-        ids: number[];
-        result: FilterSetFullData[];
-      }
-    >({
-      method: 'GET',
-      endpoint: `/api/v1/dashboard/${dashboardId}/filtersets`,
-    })(null);
-  // }
-  // return API_HANDLER.SupersetClient({
-  //   method: 'get',
-  //   url: `/api/v1/dashboard/${dashboardId}/filtersets`,
-  // });
+  const fetchFilterSets = async () => {
+    if (isStandalone) {
+      return makeApi<
+        null,
+        {
+          count: number;
+          ids: number[];
+          result: FilterSetFullData[];
+        }
+      >({
+        method: 'GET',
+        endpoint: `/api/v1/dashboard/${dashboardId}/filtersets`,
+      })(null);
+    }
+    // DODO added 44611022
+    return API_HANDLER.SupersetClient({
+      method: 'get',
+      url: `/api/v1/dashboard/${dashboardId}/filtersets`,
+    });
+  };
+
   dispatch({
     type: SET_FILTER_SETS_BEGIN,
   });
@@ -319,25 +325,28 @@ export const createFilterSet =
     const dashboardId = getState().dashboardInfo.id;
     const postFilterSets = (
       data: Partial<FilterSetFullData & { json_metadata: any }>,
-    ) =>
-      // if (process.env.type === undefined) {
-      makeApi<
-        Partial<FilterSetFullData & { json_metadata: any }>,
-        {
-          count: number;
-          ids: number[];
-          result: FilterSetFullData[];
-        }
-      >({
-        method: 'POST',
-        endpoint: `/api/v1/dashboard/${dashboardId}/filtersets`,
-      })(data);
-    // }
-    // return API_HANDLER.SupersetClient({
-    //   method: 'post',
-    //   url: `/api/v1/dashboard/${dashboardId}/filtersets`,
-    //   body: data,
-    // });
+    ) => {
+      if (isStandalone) {
+        return makeApi<
+          Partial<FilterSetFullData & { json_metadata: any }>,
+          {
+            count: number;
+            ids: number[];
+            result: FilterSetFullData[];
+          }
+        >({
+          method: 'POST',
+          endpoint: `/api/v1/dashboard/${dashboardId}/filtersets`,
+        })(data);
+      }
+      // DODO added 44611022
+      return API_HANDLER.SupersetClient({
+        method: 'post',
+        url: `/api/v1/dashboard/${dashboardId}/filtersets`,
+        body: data,
+      });
+    };
+
     dispatch({
       type: CREATE_FILTER_SET_BEGIN,
       id: -1, // -1 means the new filter set is being created
@@ -373,18 +382,22 @@ export const updateFilterSet =
     const dashboardId = getState().dashboardInfo.id;
     const postFilterSets = (
       data: Partial<FilterSetFullData & { json_metadata: any }>,
-    ) =>
-      // if (process.env.type === undefined) {
-      makeApi<Partial<FilterSetFullData & { json_metadata: any }>, {}>({
-        method: 'PUT',
-        endpoint: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSet.id}`,
-      })(data);
-    // }
-    // return API_HANDLER.SupersetClient({
-    //   method: 'put',
-    //   url: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSet.id}`,
-    //   body: data,
-    // });
+    ) => {
+      if (isStandalone) {
+        return makeApi<Partial<FilterSetFullData & { json_metadata: any }>, {}>(
+          {
+            method: 'PUT',
+            endpoint: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSet.id}`,
+          },
+        )(data);
+      }
+      // DODO added 44611022
+      return API_HANDLER.SupersetClient({
+        method: 'put',
+        url: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSet.id}`,
+        body: data,
+      });
+    };
     dispatch({
       type: UPDATE_FILTER_SET_BEGIN,
       id: filterSet.id,
@@ -418,17 +431,20 @@ export const deleteFilterSet =
   (filterSetId: number) =>
   async (dispatch: Function, getState: () => RootState) => {
     const dashboardId = getState().dashboardInfo.id;
-    const deleteFilterSets = () =>
-      // if (process.env.type === undefined) {
-      makeApi<{}, {}>({
-        method: 'DELETE',
-        endpoint: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSetId}`,
-      })({});
-    // }
-    // return API_HANDLER.SupersetClient({
-    //   method: 'delete',
-    //   url: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSetId}`,
-    // });
+    const deleteFilterSets = () => {
+      if (isStandalone) {
+        return makeApi<{}, {}>({
+          method: 'DELETE',
+          endpoint: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSetId}`,
+        })({});
+      }
+      // DODO added 44611022
+      return API_HANDLER.SupersetClient({
+        method: 'delete',
+        url: `/api/v1/dashboard/${dashboardId}/filtersets/${filterSetId}`,
+      });
+    };
+
     dispatch({
       type: DELETE_FILTER_SET_BEGIN,
       id: filterSetId,
